requirements.txt
pandas==0.24.1
tensorflow_gpu==1.13.1
opencv_python==4.1.0.25
Keras==2.2.4
scikit_image==0.17.2
numpy==1.17.0
Pillow==7.2.0
skimage==0.0
tensorflow==2.2.0